# swagger/docs/swagger.yaml
definitions:
  dao.ActivityAndUserInfo:
    properties:
      activity_id:
        description: 活动id
        type: integer
      activity_name:
        description: 活动名称
        type: string
      birthday:
        description: 出生日期
        type: string
      create_time:
        description: 创建时间
        type: string
      declare_type:
        description: 1:幼教、2:小学、3:初中、4:高中、5:职校、6:教研
        type: integer
      final_score:
        description: 最终得分（各项通过的审核）
        type: number
      identity_card:
        description: 身份证号
        type: string
      phone:
        description: 手机号
        type: string
      prize:
        description: 奖金
        type: integer
      rank:
        description: 排名
        type: integer
      rank_prize:
        description: 0：无，1：一等奖，2：二等奖，3：三等奖
        type: integer
      school_id:
        description: 学校id
        type: string
      school_name:
        description: 学校名称
        type: string
      subject_code:
        description: 科目code
        type: string
      user_activity_id:
        description: 主键id
        type: integer
      user_id:
        description: 用户id
        type: string
      user_name:
        description: 用户姓名
        type: string
      user_sex:
        description: 1：男，2：女
        type: integer
      year:
        description: 年份
        type: integer
    type: object
...
  dao.EdbReviewItem:
    properties:
      create_time:
        description: 创建时间
        type: string
      declare_type:
        description: 1:幼教、2:小学、3:初中、4:高中、5:职校、6:教研
        type: integer
      final_score:
        description: 最终得分（各项通过的审核）
        type: number
      identity_card:
        description: 身份证号
        type: string
      school_id:
        description: 学校id
        type: string
      school_name:
        description: 学校名称
        type: string
      status:
        description: 审核状态:1:待审核，2：已审核
        type: integer
      subject_code:
        description: 科目code
        type: string
      user_activity_id:
        description: 用户活动id
        type: integer
      user_id:
        description: 用户id
        type: string
      user_name:
        description: 用户姓名
        type: string
      user_sex:
        description: 1：男，2：女
        type: integer
    type: object
...
  dao.UserActivity:
    properties:
      activity_id:
        description: 活动id
        type: integer
      birthday:
        description: 出生日期
        type: string
      create_time:
        description: 创建时间
        type: string
      declare_type:
        description: 1:幼教、2:小学、3:初中、4:高中、5:职校、6:教研
        type: integer
      final_score:
        description: 最终得分（各项通过的审核）
        type: number
      identity_card:
        description: 身份证号
        type: string
      phone:
        description: 手机号
        type: string
      prize:
        description: 奖金
        type: integer
      rank:
        description: 排名
        type: integer
      rank_prize:
        description: 0：无，1：一等奖，2：二等奖，3：三等奖
        type: integer
      school_id:
        description: 学校id
        type: string
      school_name:
        description: 学校名称
        type: string
      subject_code:
        description: 科目code
        type: string
      user_activity_id:
        description: 主键id
        type: integer
      user_id:
        description: 用户id
        type: string
      user_name:
        description: 用户姓名
        type: string
      user_sex:
        description: 1：男，2：女
        type: integer
    type: object
...
  req.BatchPassReq:
    properties:
      user_activity_indicator_ids:
        description: 用户活动申报单项目id
        items:
          type: integer
        type: array
    required:
    - user_activity_indicator_ids
    type: object
...
  req.EdbDeclareToUserReq:
    properties:
      award_date:
        description: 获奖日期
        type: string
      certificate_end_date:
        description: 证书有效期——结束时间
        type: string
      certificate_start_date:
        description: 证书有效期——开始时间
        type: string
      certificate_type:
        description: 1.证书；需要填写证书有效期，2.证明；不需要有效期
        type: integer
      certificate_url:
        description: 证书url
        type: string
      two_indicator_id:
        description: 二级指标id
        type: integer
      user_activity_id:
        description: 用户活动id
        type: integer
    required:
    - two_indicator_id
    - user_activity_id
    type: object
...
  req.UpdateTwoIndicatorIdReq:
    properties:
      award_date:
        description: 获奖日期
        type: string
      certificate_end_date:
        description: 证书有效期——结束时间
        type: string
      certificate_start_date:
        description: 证书有效期——开始时间
        type: string
      certificate_type:
        description: 1.证书；需要填写证书有效期，2.证明；不需要有效期
        type: integer
      certificate_url:
        description: 证书url
        type: string
      two_indicator_id:
        description: 二级指标id
        type: integer
      user_activity_indicator_id:
        description: 用户活动申报单项目id
        type: integer
    required:
    - two_indicator_id
    - user_activity_indicator_id
    type: object
...
  resp.GetAwardsSetListResp:
    properties:
      activity_id:
        description: 活动id
        type: integer
      activity_name:
        description: 活动名称
        type: string
      award_num:
        description: 获奖人数
        type: integer
      first_prize_num:
        description: 一等奖
        type: integer
      list:
        items:
          $ref: '#/definitions/dao.UserActivity'
        type: array
      second_prize_num:
        description: 二等奖
        type: integer
      third_prize_num:
        description: 三等奖
        type: integer
      total:
        type: integer
    type: object
...
  resp.GetEdbReviewListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/dao.EdbReviewItem'
        type: array
      total:
        type: integer
    type: object
...
  resp.GetHistoryActivityListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/dao.ActivityAndUserInfo'
        type: array
      total:
        type: integer
    type: object
...
  resp.GetReviewListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/resp.ReviewItem'
        type: array
      total:
        type: integer
    type: object
...
  resp.GetWaitReviewListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/resp.WaitReview'
        type: array
      total:
        type: integer
    type: object
...
  resp.Response:
    properties:
      code:
        type: integer
      data: {}
      error_stack: {}
      msg:
        type: string
    type: object
  resp.ReviewItem:
    properties:
      activity_id:
        description: 活动id
        type: integer
      create_time:
        description: 创建时间
        type: string
      declare_type:
        description: 1:幼教、2:小学、3:初中、4:高中、5:职校、6:教研
        type: integer
      one_indicator_id:
        description: 所属一级指标
        type: integer
      one_indicator_name:
        description: 一级指标名称
        type: string
      review_status:
        description: 审核状态 1:待审核，2：未通过，3：通过
        type: integer
      school_id:
        description: 学校id
        type: string
      school_name:
        description: 学校名称
        type: string
      score:
        description: 分值
        type: number
      status:
        description: 审核状态:1.已提交 ,2.已通过,3.学校未通过,4.专家未通过,5.教育局未通过,6.活动已结束，未审批,
        type: integer
      subject_code:
        description: 科目code
        type: string
      two_indicator_id:
        description: 二级指标id
        type: integer
      two_indicator_name:
        description: 二级指标名称
        type: string
      user_activity_indicator_id:
        description: 主键id
        type: integer
      user_id:
        description: 用户id
        type: string
      user_name:
        description: 用户姓名
        type: string
      user_sex:
        description: 1：男，2：女
        type: integer
      year:
        description: 年份
        type: integer
    type: object
...
  resp.WaitReview:
    properties:
      declare_time:
        description: 申报时间
        type: string
      one_indicator_id:
        description: 一级指标id
        type: integer
      one_indicator_name:
        description: 一级指标名称
        type: string
      two_indicator_id:
        description: 二级指标id
        type: integer
      two_indicator_name:
        description: 二级指标名称
        type: string
      user_activity_indicator_id:
        description: 用户活动申报单个项目的id
        type: integer
      user_id:
        description: 用户id
        type: string
      user_name:
        description: 用户姓名
        type: string
    type: object
...
host: localhost:7850
info:
  contact:
    email: support@swagger_controller.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a teachers-awards server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger teachers-awards API
  version: "1.0"
paths:
...
  /v1/activity/edb/review/list:
    get:
      parameters:
      - description: 用户姓名
        in: query
        name: user_name
        type: string
      - description: 1：男，2：女
        in: query
        name: user_sex
        type: integer
      - description: 科目code
        in: query
        name: subject_code
        type: string
      - description: 学校id
        in: query
        name: school_id
        type: string
      - description: 学校名称
        in: query
        name: school_name
        type: string
      - description: 审核状态 0:全部 1:待审核，2：已审核
        in: query
        name: status
        type: integer
      - description: 是否仅获取数量
        in: query
        name: only_count
        type: boolean
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetEdbReviewListResp'
              type: object
      summary: 审核列表（教育局）
      tags:
      - ReviewDeclare
...
  /v1/activity/review/awards/list:
    get:
      parameters:
      - description: 活动id
        in: query
        name: activity_id
        required: true
        type: integer
      - description: 用户姓名
        in: query
        name: user_name
        type: string
      - description: 1：男，2：女
        in: query
        name: user_sex
        type: integer
      - description: 科目code
        in: query
        name: subject_code
        type: string
      - description: 学校id
        in: query
        name: school_id
        type: string
      - description: 学校名称
        in: query
        name: school_name
        type: string
      - description: 申报类型
        in: query
        name: declare_type
        type: integer
      - description: 身份证号
        in: query
        name: identity_card
        type: string
      - description: 名次，0：无，1：一等奖，2：二等奖，3：三等奖
        in: query
        name: rank_prize
        type: integer
      - description: 最终得分（各项通过的审核）
        in: query
        name: final_score
        type: number
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetAwardsSetListResp'
              type: object
      summary: 获取奖项设置列表（教育局）
      tags:
      - ReviewDeclare
  /v1/activity/review/batch/pass:
    post:
      parameters:
      - description: body请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.BatchPassReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 批量审核通过
      tags:
      - ReviewDeclare
  /v1/activity/review/commit:
    post:
      parameters:
      - description: 用户活动申报二级指标id
        in: formData
        name: user_activity_indicator_id
        required: true
        type: integer
      - description: 0：未通过，1：通过
        in: formData
        name: is_pass
        required: true
        type: integer
      - description: 审核意见
        in: formData
        name: opinion
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 提交审核
      tags:
      - ReviewDeclare
  /v1/activity/review/commit/result:
    post:
      parameters:
      - description: 活动id
        in: formData
        name: activity_id
        required: true
        type: integer
      - description: 申报类型
        in: formData
        name: declare_type
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 提交活动评审结果
      tags:
      - ReviewDeclare
  /v1/activity/review/edb/declare/user:
    post:
      parameters:
      - description: body请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.EdbDeclareToUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 教育局给用户添加申报
      tags:
      - ReviewDeclare
  /v1/activity/review/history:
    get:
      parameters:
      - description: 用户姓名
        in: query
        name: user_name
        type: string
      - description: 1：男，2：女
        in: query
        name: user_sex
        type: integer
      - description: 科目code
        in: query
        name: subject_code
        type: string
      - description: 学校id
        in: query
        name: school_id
        type: string
      - description: 学校名称
        in: query
        name: school_name
        type: string
      - description: 年份
        in: query
        name: year
        type: integer
      - description: 申报类型
        in: query
        name: declare_type
        type: integer
      - description: 身份证号
        in: query
        name: identity_card
        type: string
      - description: 0：无，1：一等奖，2：二等奖，3：三等奖
        in: query
        name: rank_prize
        type: integer
      - description: 排名
        in: query
        name: rank
        type: integer
      - description: 最终得分（各项通过的审核）
        in: query
        name: final_score
        type: number
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetHistoryActivityListResp'
              type: object
      summary: 获取历史活动评审结果列表
      tags:
      - ReviewDeclare
  /v1/activity/review/list:
    get:
      parameters:
      - description: 评委id
        in: query
        name: judges_id
        type: string
      - description: 用户姓名
        in: query
        name: user_name
        type: string
      - description: 1：男，2：女
        in: query
        name: user_sex
        type: integer
      - description: 科目code
        in: query
        name: subject_code
        type: string
      - description: 学校id
        in: query
        name: school_id
        type: string
      - description: 学校名称
        in: query
        name: school_name
        type: string
      - description: 年份
        in: query
        name: year
        type: integer
      - description: 一级指标名称
        in: query
        name: one_indicator_name
        type: string
      - description: 二级指标名称
        in: query
        name: two_indicator_name
        type: string
      - description: 审核状态 0:全部 1:待审核，2：未通过，3：已通过，4：已审核
        in: query
        name: review_status
        type: integer
      - description: 当前审核进程，1：学校，2：专家，3：教育局
        in: query
        name: review_process
        required: true
        type: integer
      - description: 是否仅获取数量
        in: query
        name: only_count
        type: boolean
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetReviewListResp'
              type: object
      summary: 审核列表
      tags:
      - ReviewDeclare
  /v1/activity/review/set/awards:
    post:
      parameters:
      - description: 活动id
        in: formData
        name: activity_id
        required: true
        type: integer
      - description: 用户id
        in: formData
        name: user_id
        required: true
        type: string
      - description: 名次
        in: formData
        name: rank_prize
        required: true
        type: integer
      - description: 奖金
        in: formData
        name: prize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 名次评定（设置奖项）
      tags:
      - ReviewDeclare
  /v1/activity/review/update/two/id:
    put:
      parameters:
      - description: body请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.UpdateTwoIndicatorIdReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 更新二级指标id
      tags:
      - ReviewDeclare
  /v1/activity/review/wait/list:
    get:
      parameters:
      - description: 当前审核进程，1：学校，2：专家，3：教育局，4：结束
        in: query
        name: review_process
        required: true
        type: integer
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetWaitReviewListResp'
              type: object
      summary: 获取待办事项列表
      tags:
      - ReviewDeclare
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"